server/requirements.txt
fastapi == 0.100.0
uvicorn == 0.22.0
hupper == 1.12
pymysql == 1.1.0
python-dotenv == 1.0.0
sqlalchemy >= 1.4.42, <1.5
alembic >= 1.3.0
passlib == 1.7.4
pydantic == 2.1.1
python-multipart == 0.0.6
python-jose == 3.3.0
botocore == 1.31.21
boto3 == 1.28.21
cryptography == 41.0.3
databases == 0.8.0
aiomysql == 0.2.0